refactor(architecture): move grammar into extension core
