+++ src/core/refactor/move_items.rs
@@ -19,7 +19,7 @@ use crate::extension::{
     self, AdjustedItem, ExtensionManifest, ParsedItem, RelatedTests, ResolvedImports,
 };
 use crate::utils::codebase_scan::{self, ExtensionFilter, ScanConfig};
-use crate::utils::grammar_items;
+use crate::extension::grammar_items;
 use crate::{component, Result};
 
 /// Result of a move operation.
