+++ src/core/scaffold/test.rs
@@ -15,8 +15,9 @@ use regex::Regex;
 use serde::Serialize;
 use std::path::{Path, PathBuf};
 
+use crate::extension::grammar;
 use crate::error::{Error, Result};
-use crate::utils::{grammar, io};
+use crate::utils::io;
 
 /// A public method/function extracted from a source file.
 #[derive(Debug, Clone, Serialize)]
